requirements/base.txt
Django==5.2.11
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
psycopg[binary]==3.3.2
dj-database-url==3.1.0
python-decouple==3.8
Pillow==12.1.1
django-storages==1.14.6
boto3==1.42.46
uvicorn[standard]==0.34.0
whitenoise==6.9.0
# Concierge v1 (per plan section 1.10, redis version adjusted for celery compat)
django-filter>=25.2
django-celery-beat>=2.8.1
django-celery-results>=2.6.0
celery[redis]>=5.6.2
django-ratelimit>=4.1.0
qrcode[pil]>=8.2
pywebpush>=2.3.0
py-vapid>=1.9.4
requests>=2.32.5
bleach>=6.0
